Norm trans_env.c adn add file with functions used by trans_env.c
